refactor: move CountrySelector para src/components e melhora funcionalidade de pesquisa da lista de paises
